move Expression/Continuation to expr.rs and cont.rs, respectively
